Moving files out (headers)
