+++ include/goofit/PDFs/physics/AmpCalc.h
@@ -1,0 +1,21 @@
+#pragma once
+
+#include <thrust/tuple.h>
+
+namespace GooFit {
+
+class AmpCalc : public thrust::unary_function<unsigned int, fpcomplex> {
+  public:
+    AmpCalc(unsigned int nPerm, unsigned int ampIdx);
+    // void setpIdx(unsigned int pIdx){_parameters = pIdx;}
+    void setDalitzId(int idx) { dalitzFuncId = idx; }
+    void setAmplitudeId(int idx) { _AmpIdx = idx; }
+    __device__ fpcomplex operator()(thrust::tuple<int, fptype *, int> t) const;
+
+  private:
+    unsigned int dalitzFuncId;
+    unsigned int _nPerm;
+    unsigned int _AmpIdx;
+};
+
+} // namespace GooFit
