+++ include/goofit/PDFs/physics/LSCalculator.h
@@ -1,0 +1,20 @@
+#pragma once
+
+#include <thrust/tuple.h>
+
+namespace GooFit {
+
+class LSCalculator : public thrust::unary_function<thrust::tuple<int, fptype *, int>, fpcomplex> {
+  public:
+    // Used to create the cached BW values.
+    LSCalculator();
+    void setDalitzId(int idx) { dalitzFuncId = idx; }
+    void setResonanceId(int idx) { _resonance_i = idx; }
+    __device__ fpcomplex operator()(thrust::tuple<int, fptype *, int> t) const;
+
+  private:
+    unsigned int dalitzFuncId;
+    unsigned int _resonance_i{0};
+};
+
+} // namespace GooFit
