+++ src/PDFs/physics/Amp4BodyT.cu
@@ -30,6 +30,13 @@ class.
 #include <mcbooster/Generate.h>
 #include <mcbooster/Vector4R.h>
 
+#include <goofit/PDFs/physics/AmpCalc_TD.h>
+#include <goofit/PDFs/physics/FourDblTupleAdd.h>
+#include <goofit/PDFs/physics/LSCalculator_TD.h>
+#include <goofit/PDFs/physics/NormIntegrator_TD.h>
+#include <goofit/PDFs/physics/NormLSCalculator_TD.h>
+#include <goofit/PDFs/physics/SFCalculator_TD.h>
+
 #include <cstdarg>
 
 namespace GooFit {
@@ -984,56 +991,6 @@ __device__ fpcomplex SFCalculator_TD::operator()(thrust::tuple<int, fptype *, in
     fptype sf = (*func)(vecs, pc);
     // printf("SpinFactors %i : %.7g\n",_spinfactor_i, sf );
     return {sf, 0.0};
-}
-
-NormSpinCalculator_TD::NormSpinCalculator_TD() = default;
-
-__device__ fptype NormSpinCalculator_TD::operator()(
-    thrust::tuple<mcbooster::GReal_t, mcbooster::GReal_t, mcbooster::GReal_t, mcbooster::GReal_t, mcbooster::GReal_t> t)
-    const {
-    // unsigned int *indices = paramIndices + _parameters; // Jump to DALITZPLOT position within parameters array
-    // int parameter_i       = 12 + (2 * indices[6]) + (indices[3] * 2)
-    //                  + (_spinfactor_i * 2); // Find position of this resonance relative to DALITZPLOT start
-    // unsigned int functn_i = indices[parameter_i];
-    // unsigned int params_i = indices[parameter_i + 1];
-
-    fptype m12   = (thrust::get<0>(t));
-    fptype m34   = (thrust::get<1>(t));
-    fptype cos12 = (thrust::get<2>(t));
-    fptype cos34 = (thrust::get<3>(t));
-    fptype phi   = (thrust::get<4>(t));
-
-    ParameterContainer pc;
-
-    fptype M  = pc.getConstant(1);
-    fptype m1 = pc.getConstant(2);
-    fptype m2 = pc.getConstant(3);
-    fptype m3 = pc.getConstant(4);
-    fptype m4 = pc.getConstant(5);
-
-    // Increment to TDDP function:
-    while(pc.funcIdx < dalitzFuncId)
-        pc.incrementIndex();
-
-    fptype vecs[16];
-    get4Vecs(vecs, m12, m34, cos12, cos34, phi, M, m1, m2, m3, m4);
-
-    //   printf("evt %i vec%i %.5g, %.5g, %.5g, %.5g\n", evtNum,0, vecs[0], vecs[1], vecs[2], vecs[3]);
-    //   printf("evt %i vec%i %.5g, %.5g, %.5g, %.5g\n", evtNum,1, vecs[4], vecs[5], vecs[6], vecs[7]);
-    //   printf("evt %i vec%i %.5g, %.5g, %.5g, %.5g\n", evtNum,2, vecs[8], vecs[9], vecs[10], vecs[11]);
-    //   printf("evt %i vec%i %.5g, %.5g, %.5g, %.5g\n", evtNum,3, vecs[12], vecs[13], vecs[14], vecs[15]);
-    // // }
-
-    while(pc.funcIdx < _spinfactor_i)
-        pc.incrementIndex();
-
-    auto func = reinterpret_cast<spin_function_ptr>(d_function_table[pc.funcIdx]);
-    fptype sf = (*func)(vecs, pc);
-
-    // printf("NormSF evt:%.5g, %.5g, %.5g, %.5g, %.5g\n", m12, m34, cos12, cos34, phi);
-    // printf("NormSF %i, %.7g\n",_spinfactor_i, sf );
-    // THREAD_SYNCH
-    return sf;
 }
 
 LSCalculator_TD::LSCalculator_TD() = default;
