+++ src/PDFs/physics/NormSpinCalculator.cu
@@ -1,0 +1,50 @@
+#include <goofit/PDFs/physics/NormSpinCalculator.h>
+
+#include <goofit/PDFs/ParameterContainer.h>
+#include <goofit/PDFs/physics/EvalVar.h>
+#include <goofit/PDFs/physics/SpinFactors.h>
+
+namespace GooFit {
+
+NormSpinCalculator::NormSpinCalculator() = default;
+
+__device__ fptype NormSpinCalculator::operator()(
+    thrust::tuple<mcbooster::GReal_t, mcbooster::GReal_t, mcbooster::GReal_t, mcbooster::GReal_t, mcbooster::GReal_t> t)
+    const {
+    fptype m12   = (thrust::get<0>(t));
+    fptype m34   = (thrust::get<1>(t));
+    fptype cos12 = (thrust::get<2>(t));
+    fptype cos34 = (thrust::get<3>(t));
+    fptype phi   = (thrust::get<4>(t));
+
+    ParameterContainer pc;
+
+    while(pc.funcIdx < dalitzFuncId)
+        pc.incrementIndex();
+
+    fptype M  = pc.getConstant(0);
+    fptype m1 = pc.getConstant(1);
+    fptype m2 = pc.getConstant(2);
+    fptype m3 = pc.getConstant(3);
+    fptype m4 = pc.getConstant(4);
+
+    fptype vecs[16];
+    get4Vecs(vecs, m12, m34, cos12, cos34, phi, M, m1, m2, m3, m4);
+
+    //   printf("evt %i vec%i %.5g, %.5g, %.5g, %.5g\n", evtNum,0, vecs[0], vecs[1], vecs[2], vecs[3]);
+    //   printf("evt %i vec%i %.5g, %.5g, %.5g, %.5g\n", evtNum,1, vecs[4], vecs[5], vecs[6], vecs[7]);
+    //   printf("evt %i vec%i %.5g, %.5g, %.5g, %.5g\n", evtNum,2, vecs[8], vecs[9], vecs[10], vecs[11]);
+    //   printf("evt %i vec%i %.5g, %.5g, %.5g, %.5g\n", evtNum,3, vecs[12], vecs[13], vecs[14], vecs[15]);
+    // // }
+    while(pc.funcIdx < _spinfactor_i)
+        pc.incrementIndex();
+
+    auto func = reinterpret_cast<spin_function_ptr>(d_function_table[pc.funcIdx]);
+    fptype sf = (*func)(vecs, pc);
+
+    // printf("NormSF evt:%.5g, %.5g, %.5g, %.5g, %.5g\n", m12, m34, cos12, cos34, phi);
+    // printf("NormSF %i, %.5g\n",_spinfactor_i, sf );
+    return sf;
+}
+
+} // namespace GooFit
